+++ example/lib/main.fpState.dart
@@ -6,98 +6,53 @@ part of 'main.dart';
 // FpStateGenerator
 // **************************************************************************
 
-extension FPAsyncState<T> on AsyncState<T> {
-  R match<R>({
-    required R Function(AsyncLoading<T> data) asyncLoading,
-    required R Function(AsyncLoaded<T> data) asyncLoaded,
-  }) {
-    final r = switch (this) {
-      AsyncLoading<T>() => asyncLoading(this as AsyncLoading<T>),
-      AsyncLoaded<T>() => asyncLoaded(this as AsyncLoaded<T>),
-      AsyncState() => throw Exception("$runtimeType not match"),
-    };
-    return r;
-  }
-
-  R matchOrElse<R>({
-    R Function(AsyncLoading<T> data)? asyncLoading,
-    R Function(AsyncLoaded<T> data)? asyncLoaded,
-    required R Function(AsyncState data) orElse,
-  }) {
-    final r = switch (this) {
-      AsyncLoading<T>() => asyncLoading == null
-          ? orElse(this as AsyncLoading<T>)
-          : asyncLoading(this as AsyncLoading<T>),
-      AsyncLoaded<T>() => asyncLoaded == null
-          ? orElse(this as AsyncLoaded<T>)
-          : asyncLoaded(this as AsyncLoaded<T>),
-      _ => orElse(this),
-    };
-    return r;
-  }
-
-  R? maybeMatch<R>({
-    R Function(AsyncLoading<T> data)? asyncLoading,
-    R Function(AsyncLoaded<T> data)? asyncLoaded,
-  }) {
-    final r = switch (this) {
-      AsyncLoading<T>() => asyncLoading?.call(this as AsyncLoading<T>),
-      AsyncLoaded<T>() => asyncLoaded?.call(this as AsyncLoaded<T>),
-      _ => throw Exception("$runtimeType not match"),
-    };
-    return r;
-  }
-}
-
 extension FPUiState on UiState {
   R match<R>({
-    required R Function(LState data) lState,
-    required R Function(RState data) rState,
-    required R Function(R3State data) r3State,
-    required R Function(R4State data) r4State,
+    required R Function() lState,
+    required R Function() rState,
+    required R Function() r3State,
+    required R Function(MemberState data) memberState,
   }) {
     final r = switch (this) {
-      LState() => lState(this as LState),
-      RState() => rState(this as RState),
-      R3State() => r3State(this as R3State),
-      R4State() => r4State(this as R4State),
+      LState() => lState(),
+      RState() => rState(),
+      R3State() => r3State(),
+      MemberState() => memberState(this as MemberState),
       UiState() => throw Exception("$runtimeType not match"),
     };
     return r;
   }
 
   R matchOrElse<R>({
-    R Function(LState data)? lState,
-    R Function(RState data)? rState,
-    R Function(R3State data)? r3State,
-    R Function(R4State data)? r4State,
+    R Function()? lState,
+    R Function()? rState,
+    R Function()? r3State,
+    R Function(MemberState data)? memberState,
     required R Function(UiState data) orElse,
   }) {
     final r = switch (this) {
-      LState() =>
-        lState == null ? orElse(this as LState) : lState(this as LState),
-      RState() =>
-        rState == null ? orElse(this as RState) : rState(this as RState),
-      R3State() =>
-        r3State == null ? orElse(this as R3State) : r3State(this as R3State),
-      R4State() =>
-        r4State == null ? orElse(this as R4State) : r4State(this as R4State),
+      LState() => lState == null ? orElse(this as LState) : lState(),
+      RState() => rState == null ? orElse(this as RState) : rState(),
+      R3State() => r3State == null ? orElse(this as R3State) : r3State(),
+      MemberState() => memberState == null
+          ? orElse(this as MemberState)
+          : memberState(this as MemberState),
       _ => orElse(this),
     };
     return r;
   }
 
   R? maybeMatch<R>({
-    R Function(LState data)? lState,
-    R Function(RState data)? rState,
-    R Function(R3State data)? r3State,
-    R Function(R4State data)? r4State,
+    R Function()? lState,
+    R Function()? rState,
+    R Function()? r3State,
+    R Function(MemberState data)? memberState,
   }) {
     final r = switch (this) {
-      LState() => lState?.call(this as LState),
-      RState() => rState?.call(this as RState),
-      R3State() => r3State?.call(this as R3State),
-      R4State() => r4State?.call(this as R4State),
+      LState() => lState?.call(),
+      RState() => rState?.call(),
+      R3State() => r3State?.call(),
+      MemberState() => memberState?.call(this as MemberState),
       _ => throw Exception("$runtimeType not match"),
     };
     return r;
@@ -106,11 +61,13 @@ extension FPUiState on UiState {
 
 extension FPFreezedState on FreezedState {
   R match<R>({
+    required R Function() noMember,
     required R Function(FreezedState1 data) freezedState1,
     required R Function(FreezedState2 data) freezedState2,
     required R Function(FreezedState3 data) freezedState3,
   }) {
     final r = switch (this) {
+      NoMember() => noMember(),
       FreezedState1() => freezedState1(this as FreezedState1),
       FreezedState2() => freezedState2(this as FreezedState2),
       FreezedState3() => freezedState3(this as FreezedState3),
@@ -120,12 +77,14 @@ extension FPFreezedState on FreezedState {
   }
 
   R matchOrElse<R>({
+    R Function()? noMember,
     R Function(FreezedState1 data)? freezedState1,
     R Function(FreezedState2 data)? freezedState2,
     R Function(FreezedState3 data)? freezedState3,
     required R Function(FreezedState data) orElse,
   }) {
     final r = switch (this) {
+      NoMember() => noMember == null ? orElse(this as NoMember) : noMember(),
       FreezedState1() => freezedState1 == null
           ? orElse(this as FreezedState1)
           : freezedState1(this as FreezedState1),
@@ -141,11 +100,13 @@ extension FPFreezedState on FreezedState {
   }
 
   R? maybeMatch<R>({
+    R Function()? noMember,
     R Function(FreezedState1 data)? freezedState1,
     R Function(FreezedState2 data)? freezedState2,
     R Function(FreezedState3 data)? freezedState3,
   }) {
     final r = switch (this) {
+      NoMember() => noMember?.call(),
       FreezedState1() => freezedState1?.call(this as FreezedState1),
       FreezedState2() => freezedState2?.call(this as FreezedState2),
       FreezedState3() => freezedState3?.call(this as FreezedState3),